baikal: Add MIPS p5600 tune file
Since Baikal-T(1) is based on ImagTech MIPS P5600 Warrior core,
we don't really need more comprehensive mips32r5 tuning. Instead
we used it as a basic tune-file for pure p5600 features declaration.
As a result the target toolchain and SDK are compatible with
P5600-based chips onle. This is what we really need, since they
are supplied to customers together with Baikal-T-based devices.

Signed-off-by: Sergey Semin <fancer.lancer@gmail.com>