machine: Add BFK/BFKx/BFK3 machine configs
These boards are produced by Baikal Electronics as the evaluation
products. Since their essence is the same as T-platform' ones,
we can freely add them to the Yocto build system without even having
actual hardware. Kernel and U-boot repos are fully based on the BE
sources, so the Yocto-based built image shall work out of the box.

Signed-off-by: Sergey Semin <fancer.lancer@gmail.com>