machine: Add Axitech v1 SFBT1/MSBT2 machine configs
This board is produced by T-platforms for Axitech company and
is also called an ADC-DAC board with serial interfaces. It can be
equipped either with SFBT1 or MSBT2 Baikal-T1 SoC modue. Kernel
and U-boot repos have already been modified to support this board,
so we can freely add the corresponding dtb and defconfig names.

Signed-off-by: Sergey Semin <fancer.lancer@gmail.com>