uboot: Add U-boot paged menu commands to the env-class
It is useful to have a two paged manu at some device configuration.
It especially concerns the ones with auto-boot and a hidden boot-menu,
which can be opened by a magic key. In this case main page would have
an ordinary device boot process and the another page would present
a recovery set of boot options. This patch addes a corresponding
functionality to the uboot-env.bbclass. No such a page selector
can be create by using "menu=main" or "menu=recovery" menuitem
commands.

Signed-off-by: Sergey Semin <fancer.lancer@gmail.com>