machine: Add Axitech BC3BT1 MRBT1 machine config
This board is produced by T-platforms for Axitech company and is in
fact very far diverged version of Axitech v1 board. It can only be
equipped with MRBT1 mezzanine card. Kernel and U-boot repos
have already been modified to support this board, so we can freely
add corresponding dtb and defconfig names.

Signed-off-by: Sergey Semin <fancer.lancer@gmail.com>