kernel: Create rsa-keys if requested
In case if U-boot signing is enabled we need rsa-keys to be available
so to properly sign the FIT-image up. If an external key directory isn't
provided lets generate them by means of openssl native utility. The output
of do_create_keys() task shall be used on the FIT-image assemly stage.

Signed-off-by: Sergey Semin <fancer.lancer@gmail.com>