yocto: Add externalsrc class usage
In case of local debugging it is necessary to supply a custom kernel/bootloader
to be built and embedded into the firmware image. In order to achieve this
linux-baikal/u-boot-baikal recipes must have EXTERNALSRC variable properly
as well as must inherit externalsrc bb-class. This commit provides the
corresponding configs. Additionally an EXTERNAL_KERNEL_SRC and EXTERNAL_UBOOT_SRC
environment variables can be used to define a path to the kernel and u-boot
respectively. They must also be added to BB_ENV_EXTRAWHITE environment
variable.

Signed-off-by: Sergey Semin <fancer.lancer@gmail.com>