kernel: Add new CNC kernel device tree files
Since the we split up the rev.3 and older devices by means of dts-files
we need to have it reflacted in the meta-baikal-t machine configs. So
we just added the absent tplatforms_cncin3-*.dtb, tplatforms_cncout3-*.dtb,
tplatforms_cncttl3-*.dtb and tplatforms_cncdac3-*.dtb files to the
KERNEL_DEVICETREE variable.

Signed-off-by: Serge Semin <fancer.lancer@gmail.com>