uboot-env: Don not use builtin echo with -ne flags
Ubuntu /bin/sh is a link to dash shell, which echo does not support
"-e" flag. Consequently the whole "-ne" flag gets printed into the
U-boot environment. Besides yocto dev manual recommends to avoid
bashisms in the shell environment. So it is better either not to
use the builtin echo or to use the coreutils echo or replace it
with printf. We decided to follow the last way.

Signed-off-by: Vadim V. Vlasov <vadim.vlasov@t-platforms.ru>