kernel: Add a recipe for kernel 4.9
Kernel version 4.9 is preferred for bn1bt1 (notebook) target.
The recipes for 4.4 and 4.9 do not differ much. Due to this we can
just utilize the kernel 4.4 recipe, but with a kernel 4.9 specific
branch for now. A dedicated recipe with kernel 4.9 libc-headers
needs to be created too, since they most likely differ from the
4.4 ones.

Since from now there are two Baikal-T charged kernels supported
we need to unpin the common recipe configs into a dedicated inc-
file.

Signed-off-by: Vadim V. Vlasov <vadim.vlasov@t-platforms.ru>