uboot-env: Add fdt U-boot menu setter command
Some mother-boards which can be equipped with T-platforms mezzanine
cards aren't always able to be runtime detected. In this case their
devices need to be manually enable just be loading a proper dtb-file.
By means of this commit it can be done via selecting a corresponding
U-boot menu, which is preliminary defined in the
UBOOT_ENV_BOOTUNITS/UBOOT_ENV_MENU_ lists.

Signed-off-by: Serge Semin <fancer.lancer@gmail.com>