uboot-env: Use load command if non or "any" fs specified
It turns out U-boot provides the 'load' command to get a file from any
filesystem. No matter whether it's ext or fat as long as it's
supported by U-boot a file will be loaded to the specified address.
Lets use it then if no particular filesystem is specified in a U-boot
environment boot menu item.

Signed-off-by: Serge Semin <fancer.lancer@gmail.com>